Replace 'sg setup' with 'sg setup2'
