+++ dev/sg/main.go
@@ -65,7 +65,6 @@ var (
 			versionCommand,
 			secretCommand,
 			setupCommand,
-			setup2Command,
 		},
 	}
 )
